api: support transposing instruments in part data (#116)
This adds an optional field to PartData named transposition. This allows you to
specify an instruments transposition at the start of the score. MusicXML encodes
this in the measure data so that you can change transpositions throughout a
part. This api implementation only allows you to set the transposition at the
start of the score (the most common use case). Subsequent instrument and
transposition changes are currently unsupported in the api.